backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.2.0  # Para manipulação de arquivos de imagem 